+++ app/src/main/java/com/billdoerr/android/carputer/SSHFragment.java
@@ -8,7 +8,11 @@ import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 
 import android.os.Handler;
+import android.util.Log;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.inputmethod.InputMethodManager;
@@ -25,12 +29,12 @@ import java.util.Calendar;
 import java.util.List;
 
 import com.billdoerr.android.carputer.settings.Node;
-import com.billdoerr.android.carputer.taskutils.AsyncTaskResponse;
-import com.billdoerr.android.carputer.taskutils.ExecuteCommandTask;
-import com.billdoerr.android.carputer.taskutils.ExecutePingTask;
-import com.billdoerr.android.carputer.taskutils.TaskCanceler;
-import com.billdoerr.android.carputer.taskutils.TaskRequest;
-import com.billdoerr.android.carputer.taskutils.TaskResult;
+import com.billdoerr.android.carputer.asynctaskutils.TaskResponse;
+import com.billdoerr.android.carputer.asynctaskutils.ExecuteCommandTask;
+import com.billdoerr.android.carputer.asynctaskutils.ExecutePingTask;
+import com.billdoerr.android.carputer.asynctaskutils.TaskTimeout;
+import com.billdoerr.android.carputer.asynctaskutils.TaskRequest;
+import com.billdoerr.android.carputer.asynctaskutils.TaskResult;
 import com.billdoerr.android.carputer.utils.FileStorageUtils;
 import com.billdoerr.android.carputer.utils.WiFiUtils;
 
@@ -38,40 +42,37 @@ import com.billdoerr.android.carputer.utils.WiFiUtils;
  *  Child fragment of CarputerFragmentMgmt.
  *  Use to perform simple remote operations on configured nodes.
  */
-public class SSHFragment extends Fragment implements AsyncTaskResponse {
+public class SSHFragment extends Fragment implements TaskResponse {
 
     private static final String TAG = "SSHFragment";
     private static final long TIMEOUT = 30*1000;     //  Time in milliseconds
 
-    // Calling Application class (see application tag in AndroidManifest.xml)
     private GlobalVariables mGlobalVariables;
-
     private List<Node> mNodes = new ArrayList<Node>();
     private WiFiUtils mWiFiUtils;
+    private static TextView sTxtReply;
 
-    //  Widgets
-    private EditText txtExecuteCommand;
-    private TextView txtReply;
-    private Spinner spinnerNodes;
+    private ProgressDialog progressDialog;
 
-    //  Misc
-    private boolean mDateSynced = false;
+    //  AsyncTask
+    private ExecuteCommandTask mExecuteCommandTask = new ExecuteCommandTask();
+    private ExecutePingTask mExecutePingTask = new ExecutePingTask();
 
-    ProgressDialog progressDialog;
-    TaskResult taskResults;
-
-    ExecuteCommandTask mExecuteCommandTask = new ExecuteCommandTask();
-    ExecutePingTask mExecutePingTask = new ExecutePingTask();
-
+    //  Indicate syncDateAll has been called
+    private static boolean sDateSynced = false;
 
     public static SSHFragment newInstance() {
         return new SSHFragment();
     }
 
+
     @SuppressLint("StaticFieldLeak")
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+
+        //  Enable options menu
+        setHasOptionsMenu(true);
 
         //  This to set delegate/listener back to this class
         mExecuteCommandTask.delegate = this;
@@ -88,10 +89,32 @@ public class SSHFragment extends Fragment implements AsyncTaskResponse {
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, final Bundle savedInstanceState) {
 
+        EditText txtExecuteCommand;
+        Spinner spinnerNodes;
+
         View view = inflater.inflate(R.layout.fragment_ssh, container, false);
 
+        //  Restore state
+        if (savedInstanceState != null) {
+            sDateSynced = savedInstanceState.getBoolean("DateSynced");
+        }
+
         //  TextView:  Reply
-        txtReply = (TextView) view.findViewById(R.id.txtReply);
+        sTxtReply = (TextView) view.findViewById(R.id.txtReply);
+
+        // Creating adapter for spinner
+        spinnerNodes = (Spinner) view.findViewById(R.id.spinnerNodes);
+        List<String> nodes = new ArrayList<>();
+        for (int i = 0; i < mNodes.size(); i++) {
+            nodes.add(mNodes.get(i).getIp());
+        }
+        ArrayAdapter<String> dataAdapterNodes = new ArrayAdapter<String>(getActivity(), android.R.layout.simple_spinner_item, nodes);
+
+        // Drop down layout style
+        dataAdapterNodes.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
+
+        // attaching data adapter to spinner
+        spinnerNodes.setAdapter(dataAdapterNodes);
 
         //  EditText:  Execute Command
         txtExecuteCommand = (EditText) view.findViewById(R.id.txtExecuteCommand);
@@ -119,18 +142,16 @@ public class SSHFragment extends Fragment implements AsyncTaskResponse {
                 p.nodes = new ArrayList<>();
                 p.nodes.add(mNodes.get(index));     //  List<Node> will be of size one, since only pinging a single node.
                 p.cmd = "";     //  Ping does not send a command
-                p.taskName = getString(R.string.msg_executing_ping);
+                p.taskName = getString(R.string.msg_executing_ping) + getString(R.string.msg_on_node) + spinnerNodes.getSelectedItem().toString();;
 
                 // Perform ping
-//                ExecutePingTask pingTask = new ExecutePingTask();
-//                new ExecutePingTask().execute(p);
                 executePingTask(p);
             }
         });
 
-        //  Button:  Poweroff (Single)
-        Button btnPoweroffSingle = (Button) view.findViewById(R.id.btnPoweroffSingle);
-        btnPoweroffSingle.setOnClickListener(new View.OnClickListener() {
+        //  Button:  PowerOff (Single)
+        Button btnPowerOffSingle = (Button) view.findViewById(R.id.btnPoweroffSingle);
+        btnPowerOffSingle.setOnClickListener(new View.OnClickListener() {
             @SuppressLint("StaticFieldLeak")
             @Override
             public void onClick(View v) {
@@ -155,9 +176,9 @@ public class SSHFragment extends Fragment implements AsyncTaskResponse {
             }
         });
 
-        //  Button:  Poweroff (All)
-        Button btnPoweroffAll = (Button) view.findViewById(R.id.btnPoweroffAll);
-        btnPoweroffAll.setOnClickListener(new View.OnClickListener() {
+        //  Button:  PowerOff (All)
+        Button btnPowerOffAll = (Button) view.findViewById(R.id.btnPoweroffAll);
+        btnPowerOffAll.setOnClickListener(new View.OnClickListener() {
             @SuppressLint("StaticFieldLeak")
             @Override
             public void onClick(View v) {
@@ -170,13 +191,13 @@ public class SSHFragment extends Fragment implements AsyncTaskResponse {
                 updateConsoleAndSystemLog(msg);
 
                 //  All nodes
-                TaskRequest p = new TaskRequest();
-                p.nodes = mNodes;
-                p.cmd = cmd;
-                p.taskName = getString(R.string.msg_power_off_all_nodes);
+                TaskRequest request = new TaskRequest();
+                request.nodes = mNodes;
+                request.cmd = cmd;
+                request.taskName = getString(R.string.msg_power_off_all_nodes);
 
                 //  Execute command
-                executeCommandTask(p);
+                executeCommandTask(request);
             }
         });
 
@@ -201,14 +222,14 @@ public class SSHFragment extends Fragment implements AsyncTaskResponse {
 
                 //  Get node from dropdown
                 int index = spinnerNodes.getSelectedItemPosition();
-                TaskRequest p = new TaskRequest();
-                p.nodes = new ArrayList<>();
-                p.nodes.add(mNodes.get(index));
-                p.cmd = cmd;
-                p.taskName = msg;
+                TaskRequest request = new TaskRequest();
+                request.nodes = new ArrayList<>();
+                request.nodes.add(mNodes.get(index));
+                request.cmd = cmd;
+                request.taskName = msg;
 
                 //  Execute command
-                executeCommandTask(p);
+                executeCommandTask(request);
             }
         });
 
@@ -230,14 +251,14 @@ public class SSHFragment extends Fragment implements AsyncTaskResponse {
 
                 //  Get node from dropdown
                 int index = spinnerNodes.getSelectedItemPosition();
-                TaskRequest p = new TaskRequest();
-                p.nodes = new ArrayList<>();
-                p.nodes.add(mNodes.get(index));
-                p.cmd = cmd;
-                p.taskName = getString(R.string.msg_sync_date_single_node) + spinnerNodes.getSelectedItem().toString();
+                TaskRequest request = new TaskRequest();
+                request.nodes = new ArrayList<>();
+                request.nodes.add(mNodes.get(index));
+                request.cmd = cmd;
+                request.taskName = getString(R.string.msg_sync_date_single_node) + spinnerNodes.getSelectedItem().toString();
 
                 //  Execute command
-                executeCommandTask(p);
+                executeCommandTask(request);
 
             }
         });
@@ -252,36 +273,6 @@ public class SSHFragment extends Fragment implements AsyncTaskResponse {
                 syncDateAll();
             }
         });
-
-        //  Nodes
-        spinnerNodes = (Spinner) view.findViewById(R.id.spinnerNodes);
-//        spinnerNodes.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
-//            @Override
-//            public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
-////                mCurrentNodeIndex = parent.getItemAtPosition(position).toString();
-//            }
-//
-//            @Override
-//            public void onNothingSelected(AdapterView<?> parent) {
-//            }
-//        });
-
-        /*
-         *  Spinner:  Nodes
-         */
-        List<String> nodes = new ArrayList<>();
-        for (int i = 0; i < mNodes.size(); i++) {
-            nodes.add(mNodes.get(i).getIp());
-        }
-
-        // Creating adapter for spinner
-        ArrayAdapter<String> dataAdapterNodes = new ArrayAdapter<String>(getActivity(), android.R.layout.simple_spinner_item, nodes);
-
-        // Drop down layout style
-        dataAdapterNodes.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-
-        // attaching data adapter to spinner
-        spinnerNodes.setAdapter(dataAdapterNodes);
 
         //  Command history
         final Spinner spinnerCommandHistory = (Spinner) view.findViewById(R.id.spinnerCommandHistory);
@@ -298,26 +289,12 @@ public class SSHFragment extends Fragment implements AsyncTaskResponse {
             }
         });
 
-//        //  TODO:  UNDER CONSTRUCTION - SystemStatus
-//        //  Button:  System Status
-//        Button btnSystemStatus = (Button) view.findViewById(R.id.btnSystemStatus);
-//        btnSystemStatus.setOnClickListener(new View.OnClickListener() {
-//            @SuppressLint("StaticFieldLeak")
-//            @Override
-//            public void onClick(View v) {
-//                //  Display dialog
-//                FragmentManager fm = getActivity().getSupportFragmentManager();
-//                CarputerMgmtFragmentSystemStatus dialogSystemStatus = CarputerMgmtFragmentSystemStatus.newInstance(getString(R.string.fragment_system_status));
-//                dialogSystemStatus.show(fm, getString(R.string.fragment_system_status));
-//            }
-//        });
-
         /*
          *  Spinner:  Command History
          */
         List<String> commandHistory = new ArrayList<>();
         commandHistory.add("date");
-        commandHistory.add("df -h /dev/sda1");
+        commandHistory.add("df -handler /dev/sda1");
         commandHistory.add("cd /mnt/motioneye/Front; ls -l");
         commandHistory.add("cd /mnt/motioneye/Rear; ls -l");
         commandHistory.add("cd /mnt/motioneye/Rear-PiCam; ls -l");
@@ -355,25 +332,108 @@ public class SSHFragment extends Fragment implements AsyncTaskResponse {
         super.onDestroy();
     }
 
+    @Override
+    public void onCreateOptionsMenu(@NonNull Menu menu, @NonNull MenuInflater inflater) {
+        super.onCreateOptionsMenu(menu, inflater);
+        inflater.inflate(R.menu.menu_options_ssh_fragment, menu);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
+        switch (item.getItemId()) {
+
+            //  Clear console
+            case R.id.action_clear_console:
+                sTxtReply.setText("");
+            default:
+                break;
+
+        }
+
+        return false;
+    }
+
+    @Override
+    public void onSaveInstanceState(Bundle savedInstanceState) {
+        super.onSaveInstanceState(savedInstanceState);
+        savedInstanceState.putBoolean("DateSynced", sDateSynced);
+    }
+
     //  This override the implemented method from asyncTask
-//    @Override
-    public void processFinish(TaskResult taskResults){
+    public void taskFinished(TaskResult result){
         //Here you will receive the result fired from async class
         //of onPostExecute(result) method.
 
-        progressDialog.hide();
-        progressDialog = null;
+        //  Dismiss the progress dialog
+        if (progressDialog !=null) {
+            progressDialog.hide();
+            progressDialog = null;
+        }
 
-        txtReply.append("Wheeeeeeeeee!");
-        updateConsoleAndSystemLog("Wheeeeeeeeee!");
+        Log.d(TAG, getString(R.string.msg_task_finished));
+        writeTaskResult(result,getString(R.string.msg_task_finished) );
+
+    }
+
+    public void hideProgress() {
+
+        //  Dismiss the progress dialog
+        if (progressDialog !=null) {
+            progressDialog.hide();
+            progressDialog = null;
+        }
+
+        Log.d(TAG, "hideProgress");
+    }
+
+    public void showProgress() {
+        Log.d(TAG, "showProgress");
+    }
+
+    public void taskTimeout(TaskResult result) {
+        Log.d(TAG, getString(R.string.msg_task_timeout));
+        writeTaskResult(result,getString(R.string.msg_task_timeout) );
+    }
+
+    public void taskCanceled(TaskResult result) {
+
+        //  Dismiss the progress dialog
+        if (progressDialog !=null) {
+            progressDialog.hide();
+            progressDialog = null;
+        }
+
+        Log.d(TAG, getString(R.string.msg_task_cancelled));
+        writeTaskResult(result,getString(R.string.msg_task_cancelled) );
+    }
+
+    private void writeTaskResult(TaskResult result, String tag) {
+
+        String msg = tag + FileStorageUtils.LINE_SEPARATOR;
+
+        if ( (result.response == null) || (result.response.isEmpty()) ) {
+            msg = msg + result.request.taskName + getString(R.string.msg_no_task_results) + FileStorageUtils.LINE_SEPARATOR;
+        } else {
+            msg = msg + result.request.taskName + getString(R.string.msg_task_results) + FileStorageUtils.LINE_SEPARATOR;
+            updateConsoleAndSystemLog(msg);
+            updateConsoleAndSystemLog(result.response);
+        }
+
+        if ( (result.exception == null) || (result.exception.isEmpty()) ) {
+            msg = msg + result.request.taskName + getString(R.string.msg_no_task_exception) + FileStorageUtils.LINE_SEPARATOR;
+        } else {
+            msg = msg + result.request.taskName + getString(R.string.msg_task_exception) + FileStorageUtils.LINE_SEPARATOR;
+            updateConsoleAndSystemLog(msg);
+            updateConsoleAndSystemLog(result.exception);
+        }
 
     }
 
     /**
-     * Wraps AsynTask in a handler to the request can be canceled after period of time.
-     * @param p TaskRequest:
+     * Wraps AsyncTask in a handler to the request can be canceled after period of time.
+     * @param request TaskRequest:
      */
-    private void executeCommandTask(TaskRequest p) {
+    private void executeCommandTask(TaskRequest request) {
 
         //  Using progress dialog even though is has been depreciated.
         progressDialog = new ProgressDialog(getActivity());
@@ -383,26 +443,28 @@ public class SSHFragment extends Fragment implements AsyncTaskResponse {
         progressDialog.show();
 
         ExecuteCommandTask task = new ExecuteCommandTask();
-        task.delegate = this;
+        task.delegate = this;   //  Callback
 
+        //  Wrap task in handler so that we can timeout task
         Handler handler = new Handler();
-        TaskCanceler taskCanceler;
-        taskCanceler = new TaskCanceler(task);
-        handler.postDelayed(taskCanceler, TIMEOUT);
+        TaskTimeout taskTimeout;
+        taskTimeout = new TaskTimeout(task);
+        handler.postDelayed(taskTimeout, TIMEOUT);
 
-        //  TODO:  I don't know if this is good technique.  Need to cancel Handler if remote command fails before timeout.
-        task.h = handler;
-        task.r = taskCanceler;
+        //  TODO:  I don't know if this is good technique.
+        //  Need to cancel Handler if remote command fails before timeout.
+        task.handler = handler;
+        task.runnable = taskTimeout;
 
         //  Execute task
-        task.execute(p);
+        task.execute(request);
     }
 
     /**
      * Wraps AsynTask in a handler to the request can be canceled after period of time.
-     * @param p TaskRequest:
+     * @param request TaskRequest:
      */
-    private void executePingTask(TaskRequest p) {
+    private void executePingTask(TaskRequest request) {
 
         //  Using progress dialog even though is has been depreciated.
         progressDialog = new ProgressDialog(getActivity());
@@ -412,19 +474,21 @@ public class SSHFragment extends Fragment implements AsyncTaskResponse {
         progressDialog.show();
 
         ExecutePingTask task = new ExecutePingTask();
-        task.delegate = this;
+        task.delegate = this;   //  Callback
 
+        //  Wrap task in handler so that we can timeout task
         Handler handler = new Handler();
-        TaskCanceler taskCanceler;
-        taskCanceler = new TaskCanceler(task);
-        handler.postDelayed(taskCanceler, TIMEOUT);
+        TaskTimeout taskTimeout;
+        taskTimeout = new TaskTimeout(task);
+        handler.postDelayed(taskTimeout, TIMEOUT);
 
-        //  TODO:  I don't know if this is good technique.  Need to cancel Handler if remote command fails before timeout.
-        task.h = handler;
-        task.r = taskCanceler;
+        //  TODO:  I don't know if this is good technique.
+        //  Need to cancel Handler if remote command fails before timeout.
+        task.handler = handler;
+        task.runnable = taskTimeout;
 
         //  Execute task
-        task.execute(p);
+        task.execute(request);
     }
 
 
@@ -434,29 +498,25 @@ public class SSHFragment extends Fragment implements AsyncTaskResponse {
     private void syncDateAll() {
         final String date = getDateTime();
         final String cmd = "sudo date -s \"" + date + "\" ;date";
+        String msg = "";
 
         //  Prepare system log and console message
-        String msg = getString(R.string.msg_executing_command) + FileStorageUtils.TABS + cmd + FileStorageUtils.LINE_SEPARATOR;
+        msg = getString(R.string.msg_executing_command) + FileStorageUtils.TABS + cmd + FileStorageUtils.LINE_SEPARATOR;
         msg = msg + getString(R.string.msg_sync_date_all_nodes) + FileStorageUtils.LINE_SEPARATOR;
         updateConsoleAndSystemLog(msg);
 
         //  All nodes
-        TaskRequest p = new TaskRequest();
-        p.nodes = mNodes;
-        p.cmd = cmd;
-        p.taskName = getString(R.string.msg_sync_date_all_nodes);
-
-        //  Check if date already synced.
-        if (mDateSynced) {
-            msg = msg + "syncDateAll: Date already synced!";
-            updateConsoleAndSystemLog(msg);
-            return;
-        }
+        TaskRequest request = new TaskRequest();
+        request.nodes = mNodes;
+        request.cmd = cmd;
+        request.taskName = getString(R.string.msg_sync_date_all_nodes);
 
         //  Execute command
-        executeCommandTask(p);
+        executeCommandTask(request);
 
-        mDateSynced = true;
+        //  Indicate that date synced so that if we return to this fragment syncDateAll is not called again.
+        sDateSynced = true;
+
     }
 
     /**
@@ -470,9 +530,17 @@ public class SSHFragment extends Fragment implements AsyncTaskResponse {
 
         // Sync dates
         if (mWiFiUtils.isConnected()) {
-            msg = getString(R.string.msg_network_connected_date_sync) + FileStorageUtils.LINE_SEPARATOR;
-            updateConsoleAndSystemLog(msg);
-//            syncDateAll();
+            //  Check if syncDateAll already run on fragment create/resume.
+            if (!sDateSynced) {
+                msg = getString(R.string.msg_network_connected_date_sync) + FileStorageUtils.LINE_SEPARATOR;
+                updateConsoleAndSystemLog(msg);
+
+                //  Sync dates
+                syncDateAll();
+            } else {
+                msg = getString(R.string.msg_sync_date_already_performed) + FileStorageUtils.LINE_SEPARATOR;
+                updateConsoleAndSystemLog(msg);
+            }
         }
 
     }
@@ -499,15 +567,7 @@ public class SSHFragment extends Fragment implements AsyncTaskResponse {
         FileStorageUtils.writeSystemLog(TAG + FileStorageUtils.TABS + msg);
 
         //  Output to console
-        txtReply.append(msg + FileStorageUtils.LINE_SEPARATOR);
-    }
-
-    private void updateConsoleAndSystemLog(String tag, String msg) {
-        //  Output to system log
-        FileStorageUtils.writeSystemLog(tag + FileStorageUtils.TABS + msg);
-
-        //  Output to console
-        txtReply.append(msg + FileStorageUtils.LINE_SEPARATOR);
+        sTxtReply.append(msg + FileStorageUtils.LINE_SEPARATOR);
     }
 
 }