+++ app/src/main/java/com/billdoerr/android/carputer/asynctaskutils/TaskRequest.java
@@ -1,14 +1,10 @@
-package com.billdoerr.android.carputer.taskutils;
-
-import android.content.Context;
+package com.billdoerr.android.carputer.asynctaskutils;
 
 import com.billdoerr.android.carputer.settings.Node;
-
 import java.util.List;
 
 //  Class AsyncTask parameters
 public class TaskRequest {
-    public Context context;
     public List<Node> nodes;
     public String cmd;
     public String taskName;