+++ app/src/main/java/com/billdoerr/android/carputer/asynctaskutils/TaskResponse.java
@@ -1,0 +1,20 @@
+package com.billdoerr.android.carputer.asynctaskutils;
+
+
+/**
+ *
+ */
+public interface TaskResponse {
+
+    void taskFinished(TaskResult taskResult);
+
+    void hideProgress();
+
+    void showProgress();
+
+    void taskTimeout(TaskResult taskResult);
+
+    void taskCanceled(TaskResult taskResult);
+
+
+}