+++ app/src/main/java/com/billdoerr/android/carputer/asynctaskutils/TaskResult.java
@@ -1,7 +1,7 @@
-package com.billdoerr.android.carputer.taskutils;
+package com.billdoerr.android.carputer.asynctaskutils;
 
 public class TaskResult {
     public TaskRequest request;
-    public String cmdResult;
+    public String response;
     public String exception;
 }