+++ app/src/main/java/com/billdoerr/android/carputer/taskutils/AsyncTaskResponse.java
@@ -1,5 +1,0 @@
-package com.billdoerr.android.carputer.taskutils;
-
-public interface AsyncTaskResponse {
-    void processFinish(TaskResult taskResult);
-}